Jms funcionando e insertando en la bd

+++ practica-4-PWA/src/main/java/jms/practica4pwa/models/Message.java
@@ -19,7 +19,7 @@ public class Message {
     private Integer idDevice;
 
     @Column
-    private Float temperature;
+    private Integer temperature;
 
     @Column
     private Integer humidity;
